requirements.txt
fastapi
uvicorn[standard]
langchain>=0.2.2
langchain-google-genai>=0.1.3
google-generativeai>=0.3.1
qdrant-client>=1.8.2
python-dotenv>=1.0.1
pdfplumber>=0.10.3
pytesseract>=0.3.10
pillow>=10.0.0
ragas>=0.0.24
